backend/requirements.txt
# Core Backend
fastapi>=0.104.1
pydantic>=2.5.2
celery>=5.3.6
redis>=5.0.1
uvicorn>=0.24.0

# AI/ML Core
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
diffusers>=0.25.0
safetensors>=0.4.0

# Image Generation (Stable Diffusion)
# Already covered by diffusers

# Video Generation (Stable Video Diffusion)
# Already covered by diffusers
opencv-python-headless>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
av>=10.0.0  # For video processing

# Audio Generation
# TTS - Coqui TTS (XTTS) - Compatible with Python 3.10+
# Using git install for latest compatible version
# TTS will be installed via: pip install git+https://github.com/coqui-ai/TTS
audiocraft  # Meta MusicGen (no version constraint for compatibility)
scipy>=1.11.0

# Video Editing
# FFmpeg (install via system: apt-get install ffmpeg)

# Image Processing
Pillow>=10.0.0

# LLM Integration
openai>=1.3.0  # Works with OpenAI, Groq, and other OpenAI-compatible APIs

# Utilities
python-dotenv>=1.0.0  # For environment variables
numpy>=1.24.0
